+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - Added error message to indicate no column names are found
 - Added TRUNCATE function to table view
 - Added DROP function to table view
+- Added TRUNCATE/DROP all tables function to index view
 
 ## Version 2.2.3